Warn when undefined identifier used in preprocessor conditional
This can mask an error when the user either typos a macro name when writing a conditional, or (as was the case for the user who pointed out this issue) they mistakenly assume that a `#define` in an `import`ed file has been made visible to them.

This change just adds the warning in the obvious place, with a test code to ensure it triggers.