Fix problem with __import not working  (#688)
* Added getPathType to ISlangFileSystemExt.
This is needed so that when searching for a file it's existance can be tested without loading the file. On some platforms a getCanonicalPath can do this - but depending on how getCanonicalPath is implemented, it may not do. This test is made after the relative path is produced before finding the canonical path.

* Test for importing along search path.

* Added comment to explain the issue around WrapFileSystem impl of getPathType.

* Make search path use / not \